Mac OS Xのサウンドモジュールを分離
