+++ template/.vgs2mkpj.cfg
@@ -12,6 +12,7 @@ cpy vgs2a.c src
 cpy vgs2w.cpp src
 cpy vgs2l.c src
 cpy vgs2x.c src
+cpy vgs2xs.c src
 cpy miniz.h src
 cpy maincxx.cpp src
 cpy vgs2i.c src/iPhone
