+++ template/vgs2.h
@@ -164,6 +164,8 @@ void make_pallet();
 int gload(unsigned char n,const char* name);
 int eload(unsigned char n,const char* name);
 int bload(unsigned char n,const char* name);
+int init_sound();
+void term_sound();
 
 /*
  *----------------------------------------------------------------------------
